#580: keep a fractional HTTP timeout instead of rounding it to zero
The four transfer commands built their client with
`timeout: (opts['timeout'] || 30).to_i`, so `--timeout 0.5` arrived at
`BazaRb` as `0`. A request then had no time at all and failed
immediately, which is the opposite of what asking for half a second
means.

`to_f` keeps it. `BazaRb#initialize` takes `timeout:` and passes it
through without coercion, and both Faraday and `Timeout.timeout` accept
a Float, so nothing downstream has to change. A whole number still
behaves as before, since 30 and 30.0 are the same number of seconds.

The same line appears in push, pull, download and upload, and all four
are changed. The test covers push, which is the shape of all of them: it
stands in for `BazaRb` to read the timeout it was handed, and without
this change it reads 0 instead of 0.5.

+++ lib/judges/commands/push.rb
@@ -36,7 +36,7 @@ class Judges::Push
     baza = BazaRb.new(
       opts['host'], opts['port'].to_i, opts['token'],
       ssl: opts['ssl'],
-      timeout: (opts['timeout'] || 30).to_i,
+      timeout: (opts['timeout'] || 30).to_f,
       loog: @loog,
       retries: (opts['retries'] || 3).to_i,
       compress: opts.fetch('zip', true)
